refactor(server): extract auth middleware to middleware/auth-setup.ts (#4243 step 1)
- Move setupAuth(), rate limiting helpers, requestKeyMap to middleware/auth-setup.ts
- server.ts: 1609 → 1314 lines (-295)
- setupAuth() now takes (app, ctx) instead of implicit app closure
- getRateLimiter() exported for cleanup access
- All 5628 tests pass, zero TS errors
- No behavioral changes — pure extraction
